requirements.txt
numpy==1.23
websockets
colorama
seaborn
kornia
natsort
tqdm
wandb
scipy
pynndescent
matplotlib
lmfit
kaleido
Pillow
gradio
tensorboard
simulus
opencv-python-headless